+++ settings/lib/RuleEditor/LoanRulesEditor.js
@@ -35,11 +35,6 @@ const propTypes = {
   completionLists: PropTypes.object,
 
   /*
-    polcies that could be applied to selections of typeGroups.
-  */
-  policies: PropTypes.arrayOf(PropTypes.object),
-
-  /*
     values that could be applied to selections of typeGroups. Defaults to:
     {
       'g': 'Patron Groups',
@@ -53,6 +48,16 @@ const propTypes = {
     A value of this should be applied for each typegroup that's expected to be highlighted by the editor.
   */
   typeMapping: PropTypes.object.isRequired,
+
+  /*
+    values that could be applied to selections of policies:
+    {
+      'l': 'Loan policies',
+      'r': 'Request policies',
+      'n': 'Notice policies',
+    }
+  */
+  policyMapping: PropTypes.object.isRequired,
 
   /*
     the code that appears in the editor
@@ -196,12 +201,12 @@ class LoanRulesEditor extends React.Component {
   componentWillReceiveProps(nextProps) {
     const nextState = {};
 
-    if (nextProps.policies && !isEqual(nextProps.policies, this.props.polices)) {
-      nextState.policies = nextProps.policies;
-    }
-
     if (nextProps.typeMapping && !isEqual(nextProps.typeMapping, this.props.typeMapping)) {
       nextState.typeMapping = nextProps.typeMapping;
+    }
+
+    if (nextProps.policyMapping && !isEqual(nextProps.policyMapping, this.props.policyMapping)) {
+      nextState.policyMapping = nextProps.policyMapping;
     }
 
     if (nextProps.completionLists && !isEqual(nextProps.completionLists, this.props.completionLists)) {
@@ -229,16 +234,23 @@ class LoanRulesEditor extends React.Component {
   }
 
   getInitialState() {
+    const {
+      completionLists,
+      typeMapping,
+      policyMapping,
+    } = this.props;
+
+    const name = 'loanRulesCMM';
     const modeConfig = {
-      name: 'loanRulesCMM',
-      completionLists: this.props.completionLists,
-      policies: this.props.policies,
-      typeMapping: this.props.typeMapping,
+      name,
+      completionLists,
+      typeMapping,
+      policyMapping,
       keySelector: [
         'all',
         'rare'
-      ]
-    }
+      ],
+    };
 
     const openTri = generateOpen();
     const foldedTri = generateFolded();