+++ test/bigtest/tests/loan-rules-test.js
@@ -22,12 +22,47 @@ describe('LoanRules', () => {
 
   describe('entering loan rule into editor', () => {
     beforeEach(async function () {
-      await loanRules.whenLoaded();
-      await loanRules.editor.setValue('m book: example-loan-policy');
+      await loanRules.editor.setValue('m book: l example-loan-policy');
     });
 
     it('should set new loan rule', () => {
-      expect(loanRules.editor.value).to.be.equal('m book: example-loan-policy');
+      expect(loanRules.editor.value).to.be.equal('m book: l example-loan-policy');
+    });
+  });
+
+  describe('entering item type', () => {
+    beforeEach(async function () {
+      await loanRules.editor.setValue('m book: ');
+    });
+
+    it('should display policy types menu', () => {
+      expect(loanRules.editor.hints.arePresent).to.be.true;
+      expect(loanRules.editor.hints.text).to.be.equal('Loan policies');
+    });
+  });
+
+  describe('choosing policy type', () => {
+    beforeEach(async function () {
+      await loanRules.editor.setValue('m book: ');
+      await loanRules.editor.pickHint(1);
+    });
+
+    it('should choose policy type', () => {
+      expect(loanRules.editor.value).to.be.equal('m book: l ');
+    });
+  });
+
+  describe('choosing loan policy', () => {
+    beforeEach(async function () {
+      await loanRules.editor.setValue('m book: ');
+      // choose policy type
+      await loanRules.editor.pickHint(1);
+      // choose loan policy
+      await loanRules.editor.pickHint(0);
+    });
+
+    it('should choose loan policy', () => {
+      expect(loanRules.editor.value).to.be.equal('m book: l example-loan-policy ');
     });
   });
 });