+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.5.0 IN PROGRESS
 
+* Add new policy types to policy menu in rules editor. Fixes UICIRC-164.
+
 ## 1.4.0 (https://github.com/folio-org/ui-circulation/tree/v1.4.0) (2018-12-13)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.3.0...v1.4.0)